+++ BHoM/Environment/Elements/Building.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using BH.oM.Geometry;
 using BH.oM.Base;
-using BH.oM.Environmental.Interface;
-using BH.oM.Environmental.Properties;
+using BH.oM.Environment.Interface;
+using BH.oM.Environment.Properties;
 using BH.oM.Architecture.Elements;
 
-namespace BH.oM.Environmental.Elements
+namespace BH.oM.Environment.Elements
 {
     public class Building : BHoMObject, IBuilidng
     {
