+++ BHoM/Environment/Elements/BuildingElement.cs
@@ -1,11 +1,11 @@
 ﻿using BH.oM.Base;
-using BH.oM.Environmental.Interface;
-using BH.oM.Environmental.Properties;
+using BH.oM.Environment.Interface;
+using BH.oM.Environment.Properties;
 using BH.oM.Architecture.Elements;
 using System.Collections.Generic;
 using System;
 
-namespace BH.oM.Environmental.Elements
+namespace BH.oM.Environment.Elements
 {
     public class BuildingElement : BHoMObject, IBuildingObject
     {
