+++ BHoM/Environment/Elements/ConstructionLayer.cs
@@ -5,18 +5,18 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
-using BH.oM.Geometry;
-using BH.oM.Environmental.Interface;
+using BH.oM.Environment.Interface;
 
-namespace BH.oM.Environmental.Elements
+namespace BH.oM.Environment.Elements
 {
-    public class BuildingElementCurve : BHoMObject, IBuildingElementGeometry
+    public class ConstructionLayer: BHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public ICurve Curve { get; set; } = new Line();
+        public IMaterial Material { get; set; } = new OpaqueMaterial();
+        public double Thickness { get; set; } = 0.0;
 
         /***************************************************/
     }
