+++ BHoM/Environment/Elements/InternalCondition.cs
@@ -1,18 +1,22 @@
-﻿using BH.oM.Environmental.Properties;
-using BH.oM.Environmental.Interface;
-using BH.oM.Base;
-using BH.oM.Architecture.Elements;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace BH.oM.Environmental.Elements
+using BH.oM.Base;
+using BH.oM.Environment.Properties;
+
+namespace BH.oM.Environment.Elements
 {
-    public class Emitter : BHoMObject, IEquipment
+    public class InternalCondition : BHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
         public EmitterProperties EmitterProperties { get; set; } = new EmitterProperties();
-        public Level Level { get; set; } = new Level();
+        public List<Profile> Profiles { get; set; } = new List<Profile>();
 
         /***************************************************/
     }
