+++ BHoM/Environment/Interface/IBuilidng.cs
@@ -1,21 +1,18 @@
-﻿using System;
+﻿using BH.oM.Base;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-using BH.oM.Geometry;
-using BH.oM.Base;
-
-namespace BH.oM.Environmental.Elements
+namespace BH.oM.Environment.Interface
 {
-    public class BuildingElementOpening : BHoMObject
+    public interface IBuilidng : IObject
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        public PolyCurve PolyCurve { get; set; } = new PolyCurve();
 
 
         /***************************************************/
