+++ BHoM/Environment/Interface/IEquipment.cs
@@ -5,11 +5,12 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
+using BH.oM.Environment.Interface;
 using BH.oM.Structural.Elements;
 
-namespace BH.oM.Environmental.Interface
+namespace BH.oM.Environment.Interface
 {
-    public interface IBuildingObject : IBHoMObject
+    public interface IEquipment : IBuildingObject
     {
         /***************************************************/
         /**** Properties                                ****/
