+++ BHoM/Environment/Properties/BuildingElementProperties.cs
@@ -5,9 +5,9 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
-using BH.oM.Environmental.Elements;
+using BH.oM.Environment.Elements;
 
-namespace BH.oM.Environmental.Properties
+namespace BH.oM.Environment.Properties
 {
     public class BuildingElementProperties : BHoMObject
     {
