+++ BHoM/Environment/Properties/EmitterProperties.cs
@@ -1,13 +1,15 @@
-﻿using BH.oM.Base;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Environmental.Interface
+using BH.oM.Base;
+using BH.oM.Environment.Interface;
+
+namespace BH.oM.Environment.Properties
 {
-    public interface IBuilidng : IObject
+    public class EmitterProperties : BHoMObject, IEquipmentProperties
     {
         /***************************************************/
         /**** Properties                                ****/
